requirements.txt
python-telegram-bot==20.7
o1.api==0.1.0
python-dotenv==1.0.0